+++ simpletransformers/ner/ner_model.py
@@ -29,6 +29,9 @@ from torch.utils.data import DataLoader, RandomSampler, SequentialSampler, Tenso
 from transformers import (
     WEIGHTS_NAME,
     AdamW,
+    AutoConfig,
+    AutoTokenizer,
+    AutoModelForTokenClassification,
     BertConfig,
     BertForTokenClassification,
     BertTokenizer,
@@ -104,6 +107,13 @@ class NERModel:
         self.args = {"classification_report": False}
         self.args.update(global_args)
 
+        saved_model_args = self._load_model_args(model_name)
+        if saved_model_args:
+            self.args.update(saved_model_args)
+
+        if args:
+            self.args.update(args)
+
         if not use_cuda:
             self.args["fp16"] = False
 
@@ -111,12 +121,13 @@ class NERModel:
             self.args.update(args)
 
         MODEL_CLASSES = {
+            "auto": (AutoConfig, AutoTokenizer, AutoModelForTokenClassification),
             "bert": (BertConfig, BertForTokenClassification, BertTokenizer),
-            "roberta": (RobertaConfig, RobertaForTokenClassification, RobertaTokenizer),
-            "distilbert": (DistilBertConfig, DistilBertForTokenClassification, DistilBertTokenizer),
             "camembert": (CamembertConfig, CamembertForTokenClassification, CamembertTokenizer),
-            "xlmroberta": (XLMRobertaConfig, XLMRobertaForTokenClassification, XLMRobertaTokenizer),
+            "distilbert": (DistilBertConfig, DistilBertForTokenClassification, DistilBertTokenizer),
             "electra": (ElectraConfig, ElectraForTokenClassification, ElectraTokenizer),
+            "roberta": (RobertaConfig, RobertaForTokenClassification, RobertaTokenizer),
+            "xlmroberta": (XLMRobertaConfig, XLMRobertaForTokenClassification, XLMRobertaTokenizer),
         }
 
         config_class, model_class, tokenizer_class = MODEL_CLASSES[model_type]
@@ -165,7 +176,9 @@ class NERModel:
             warnings.warn("wandb_project specified but wandb is not available. Wandb disabled.")
             self.args["wandb_project"] = None
 
-    def train_model(self, train_data, output_dir=None, show_running_loss=True, args=None, eval_data=None, verbose=True, **kwargs):
+    def train_model(
+        self, train_data, output_dir=None, show_running_loss=True, args=None, eval_data=None, verbose=True, **kwargs
+    ):
         """
         Trains the model using 'train_data'
 
@@ -394,7 +407,9 @@ class NERModel:
                         and global_step % args["evaluate_during_training_steps"] == 0
                     ):
                         # Only evaluate when single GPU otherwise metrics may not average well
-                        results, _, _ = self.eval_model(eval_data, verbose=verbose and args["evaluate_during_training_verbose"], **kwargs)
+                        results, _, _ = self.eval_model(
+                            eval_data, verbose=verbose and args["evaluate_during_training_verbose"], **kwargs
+                        )
                         for key, value in results.items():
                             tb_writer.add_scalar("eval_{}".format(key), value, global_step)
 
@@ -485,7 +500,9 @@ class NERModel:
                 self._save_model(output_dir_current, optimizer, scheduler, model=model)
 
             if args["evaluate_during_training"]:
-                results, _, _ = self.eval_model(eval_data, verbose=verbose and args["evaluate_during_training_verbose"], **kwargs)
+                results, _, _ = self.eval_model(
+                    eval_data, verbose=verbose and args["evaluate_during_training_verbose"], **kwargs
+                )
 
                 self._save_model(output_dir_current, optimizer, scheduler, results=results)
 
@@ -571,7 +588,9 @@ class NERModel:
 
         eval_dataset = self.load_and_cache_examples(eval_data, evaluate=True)
 
-        result, model_outputs, preds_list = self.evaluate(eval_dataset, output_dir, verbose=verbose, silent=silent, **kwargs)
+        result, model_outputs, preds_list = self.evaluate(
+            eval_dataset, output_dir, verbose=verbose, silent=silent, **kwargs
+        )
         self.results.update(result)
 
         if verbose:
@@ -653,7 +672,7 @@ class NERModel:
             "precision": precision_score(out_label_list, preds_list),
             "recall": recall_score(out_label_list, preds_list),
             "f1_score": f1_score(out_label_list, preds_list),
-            **extra_metrics
+            **extra_metrics,
         }
 
         results.update(result)
@@ -681,8 +700,8 @@ class NERModel:
 
         Returns:
             preds: A Python list of lists with dicts containing each word mapped to its NER tag.
-            model_outputs: A python list of the raw model outputs for each text.
-        """
+            model_outputs: A Python list of lists with dicts containing each word mapped to its list with raw model output.
+        """  # noqa: ignore flake8"
 
         device = self.device
         model = self.model
@@ -735,12 +754,18 @@ class NERModel:
             if preds is None:
                 preds = logits.detach().cpu().numpy()
                 out_label_ids = inputs["labels"].detach().cpu().numpy()
+                out_input_ids = inputs["input_ids"].detach().cpu().numpy()
+                out_attention_mask = inputs["attention_mask"].detach().cpu().numpy()
             else:
                 preds = np.append(preds, logits.detach().cpu().numpy(), axis=0)
                 out_label_ids = np.append(out_label_ids, inputs["labels"].detach().cpu().numpy(), axis=0)
+                out_input_ids = np.append(out_input_ids, inputs["input_ids"].detach().cpu().numpy(), axis=0)
+                out_attention_mask = np.append(
+                    out_attention_mask, inputs["attention_mask"].detach().cpu().numpy(), axis=0,
+                )
 
         eval_loss = eval_loss / nb_eval_steps
-        model_outputs = preds
+        token_logits = preds
         preds = np.argmax(preds, axis=2)
 
         label_map = {i: label for i, label in enumerate(self.labels)}
@@ -759,7 +784,49 @@ class NERModel:
             for i, sentence in enumerate(to_predict)
         ]
 
+        word_tokens = []
+        for n, sentence in enumerate(to_predict):
+            w_log = self._convert_tokens_to_word_logits(
+                out_input_ids[n], out_label_ids[n], out_attention_mask[n], token_logits[n],
+            )
+            word_tokens.append(w_log)
+
+        model_outputs = [
+            [{word: word_tokens[i][j]} for j, word in enumerate(sentence.split()[: len(preds_list[i])])]
+            for i, sentence in enumerate(to_predict)
+        ]
         return preds, model_outputs
+
+    def _convert_tokens_to_word_logits(self, input_ids, label_ids, attention_mask, logits):
+
+        ignore_ids = [
+            self.tokenizer.convert_tokens_to_ids(self.tokenizer.pad_token),
+            self.tokenizer.convert_tokens_to_ids(self.tokenizer.sep_token),
+            self.tokenizer.convert_tokens_to_ids(self.tokenizer.cls_token),
+        ]
+
+        # Remove unuseful positions
+        masked_ids = input_ids[(1 == attention_mask)]
+        masked_labels = label_ids[(1 == attention_mask)]
+        masked_logits = logits[(1 == attention_mask)]
+        for id in ignore_ids:
+            masked_labels = masked_labels[(id != masked_ids)]
+            masked_logits = masked_logits[(id != masked_ids)]
+            masked_ids = masked_ids[(id != masked_ids)]
+
+        # Map to word logits
+        word_logits = []
+        tmp = []
+        for n, lab in enumerate(masked_labels):
+            if lab != self.pad_token_label_id:
+                if n != 0:
+                    word_logits.append(tmp)
+                tmp = [list(masked_logits[n])]
+            else:
+                tmp.append(list(masked_logits[n]))
+        word_logits.append(tmp)
+
+        return word_logits
 
     def load_and_cache_examples(self, data, evaluate=False, no_cache=False, to_predict=None):
         """
@@ -888,9 +955,22 @@ class NERModel:
             torch.save(self.args, os.path.join(output_dir, "training_args.bin"))
             torch.save(optimizer.state_dict(), os.path.join(output_dir, "optimizer.pt"))
             torch.save(scheduler.state_dict(), os.path.join(output_dir, "scheduler.pt"))
+            self._save_model_args(output_dir)
 
         if results:
             output_eval_file = os.path.join(output_dir, "eval_results.txt")
             with open(output_eval_file, "w") as writer:
                 for key in sorted(results.keys()):
                     writer.write("{} = {}\n".format(key, str(results[key])))
+
+    def _save_model_args(self, output_dir):
+        os.makedirs(output_dir, exist_ok=True)
+        with open(os.path.join(output_dir, "model_args.json"), "w") as f:
+            json.dump(self.args, f)
+
+    def _load_model_args(self, input_dir):
+        model_args_file = os.path.join(input_dir, "model_args.json")
+        if os.path.isfile(model_args_file):
+            with open(model_args_file, "r") as f:
+                model_args = json.load(f)
+            return model_args