+++ simpletransformers/seq2seq/seq2seq_model.py
@@ -13,12 +13,41 @@ from tqdm.auto import tqdm, trange
 import pandas as pd
 import torch
 from simpletransformers.config.global_args import global_args
-from simpletransformers.t5.t5_utils import T5Dataset
+from simpletransformers.seq2seq.seq2seq_utils import Seq2SeqDataset, SimpleSummarizationDataset
 from tensorboardX import SummaryWriter
 from torch.nn.utils.rnn import pad_sequence
 from torch.utils.data import DataLoader, Dataset, RandomSampler, SequentialSampler
 from torch.utils.data.distributed import DistributedSampler
-from transformers import AdamW, T5Config, T5ForConditionalGeneration, T5Tokenizer, get_linear_schedule_with_warmup
+from transformers import AdamW, EncoderDecoderModel, EncoderDecoderConfig, get_linear_schedule_with_warmup
+from transformers import (
+    AutoModel,
+    AutoTokenizer,
+    AutoConfig,
+    BertTokenizer,
+    BertModel,
+    BertForMaskedLM,
+    BertConfig,
+    CamembertConfig,
+    CamembertModel,
+    CamembertTokenizer,
+    DistilBertConfig,
+    DistilBertModel,
+    DistilBertTokenizer,
+    ElectraConfig,
+    ElectraModel,
+    ElectraTokenizer,
+    PreTrainedModel,
+    PreTrainedTokenizer,
+    RobertaConfig,
+    RobertaModel,
+    RobertaTokenizer,
+    BartForConditionalGeneration,
+    BartTokenizer,
+    BartConfig,
+    MarianMTModel,
+    MarianSentencePieceTokenizer,
+    MarianConfig,
+)
 
 try:
     import wandb
@@ -29,28 +58,64 @@ except ImportError:
 
 logger = logging.getLogger(__name__)
 
+MODEL_CLASSES = {
+    "auto": (AutoConfig, AutoModel, AutoTokenizer),
+    "bert": (BertConfig, BertModel, BertTokenizer),
+    "roberta": (RobertaConfig, RobertaModel, RobertaTokenizer),
+    "distilbert": (DistilBertConfig, DistilBertModel, DistilBertTokenizer),
+    "camembert": (CamembertConfig, CamembertModel, CamembertTokenizer),
+    "electra": (ElectraConfig, ElectraModel, ElectraTokenizer),
+    "bart": (BartConfig, BartForConditionalGeneration, BartTokenizer),
+    "marian": (MarianConfig, MarianMTModel, MarianSentencePieceTokenizer),
+}
 
-def chunks(lst, n):
-    """Yield successive n-sized chunks from lst."""
-    for i in range(0, len(lst), n):
-        yield lst[i : i + n]
 
-
-class T5Model:
+class Seq2SeqModel:
     def __init__(
-        self, model_name, args=None, use_cuda=True, cuda_device=-1, **kwargs,
+        self,
+        encoder_type=None,
+        encoder_name=None,
+        decoder_name=None,
+        encoder_decoder_type=None,
+        encoder_decoder_name=None,
+        config=None,
+        args=None,
+        use_cuda=True,
+        cuda_device=-1,
+        **kwargs,
     ):
 
         """
-        Initializes a T5Model model.
+        Initializes a Seq2SeqModel.
 
         Args:
-            model_name: The exact architecture and trained weights to use. This may be a Hugging Face Transformers compatible pre-trained model, a community model, or the path to a directory containing model files.
+            encoder_type (optional): The type of model to use as the encoder.
+            encoder_name (optional): The exact architecture and trained weights to use. This may be a Hugging Face Transformers compatible pre-trained model, a community model, or the path to a directory containing model files.
+            decoder_name (optional): The exact architecture and trained weights to use. This may be a Hugging Face Transformers compatible pre-trained model, a community model, or the path to a directory containing model files.
+                                    Must be the same "size" as the encoder model (base/base, large/large, etc.)
+            encoder_decoder_type (optional): The type of encoder-decoder model. (E.g. bart)
+            encoder_decoder_name (optional): The path to a directory containing the saved encoder and decoder of a Seq2SeqModel. (E.g. "outputs/") OR a valid BART model.
+            config (optional): A configuration file to build an EncoderDecoderModel.
             args (optional): Default args will be used if this parameter is not provided. If provided, it should be a dict containing the args that should be changed in the default args.
             use_cuda (optional): Use GPU if available. Setting to False will force model to use CPU only.
             cuda_device (optional): Specific GPU that should be used. Will use the first available GPU by default.
             **kwargs (optional): For providing proxies, force_download, resume_download, cache_dir and other options specific to the 'from_pretrained' implementation where this will be supplied.
         """  # noqa: ignore flake8"
+
+        if not config:
+            # if not ((encoder_name and decoder_name) or encoder_decoder_name) and not encoder_type:
+            if not ((encoder_name and decoder_name) or encoder_decoder_name):
+                raise ValueError(
+                    "You must specify a Seq2Seq config \t OR \t"
+                    "encoder_type, encoder_name, and decoder_name OR \t \t"
+                    "encoder_type and encoder_decoder_name"
+                )
+            elif not (encoder_type or encoder_decoder_type):
+                raise ValueError(
+                    "You must specify a Seq2Seq config \t OR \t"
+                    "encoder_type, encoder_name, and decoder_name \t OR \t"
+                    "encoder_type and encoder_decoder_name"
+                )
 
         if args and "manual_seed" in args:
             random.seed(args["manual_seed"])
@@ -69,17 +134,13 @@ class T5Model:
             "repetition_penalty": 1.0,
             "length_penalty": 2.0,
             "early_stopping": True,
-            "preprocess_inputs": True,
         }
 
         self.args.update(global_args)
 
-        saved_model_args = self._load_model_args(model_name)
+        saved_model_args = self._load_model_args(encoder_decoder_name)
         if saved_model_args:
             self.args.update(saved_model_args)
-
-        if args:
-            self.args.update(args)
 
         if args:
             self.args.update(args)
@@ -100,20 +161,68 @@ class T5Model:
 
         self.results = {}
 
-        self.config = T5Config.from_pretrained(model_name, **self.args["config"])
-
-        self.model = T5ForConditionalGeneration.from_pretrained(model_name, config=self.config)
-
-        self.tokenizer = T5Tokenizer.from_pretrained(model_name)
-
         if not use_cuda:
             self.args["fp16"] = False
 
-        self.args["model_name"] = model_name
+        # config = EncoderDecoderConfig.from_encoder_decoder_configs(config, config)
+        if encoder_decoder_type:
+            config_class, model_class, tokenizer_class = MODEL_CLASSES[encoder_decoder_type]
+        else:
+            config_class, model_class, tokenizer_class = MODEL_CLASSES[encoder_type]
+
+        if encoder_decoder_type in ["bart", "marian"]:
+            self.model = model_class.from_pretrained(encoder_decoder_name)
+            if encoder_decoder_type == "bart":
+                self.encoder_tokenizer = tokenizer_class.from_pretrained(encoder_decoder_name)
+            elif encoder_decoder_type == "marian":
+                if "base_marian_model_name" in self.args:
+                    self.encoder_tokenizer = tokenizer_class.from_pretrained(self.args["base_marian_model_name"])
+                else:
+                    self.encoder_tokenizer = tokenizer_class.from_pretrained(encoder_decoder_name)
+            self.decoder_tokenizer = self.encoder_tokenizer
+            self.config = self.model.config
+        else:
+            if encoder_decoder_name:
+                # self.model = EncoderDecoderModel.from_pretrained(encoder_decoder_name)
+                self.model = EncoderDecoderModel.from_encoder_decoder_pretrained(
+                    os.path.join(encoder_decoder_name, "encoder"), os.path.join(encoder_decoder_name, "decoder")
+                )
+                self.model.encoder = model_class.from_pretrained(os.path.join(encoder_decoder_name, "encoder"))
+                self.model.decoder = BertForMaskedLM.from_pretrained(os.path.join(encoder_decoder_name, "decoder"))
+                self.encoder_tokenizer = tokenizer_class.from_pretrained(os.path.join(encoder_decoder_name, "encoder"))
+                self.decoder_tokenizer = BertTokenizer.from_pretrained(os.path.join(encoder_decoder_name, "decoder"))
+            else:
+                self.model = EncoderDecoderModel.from_encoder_decoder_pretrained(
+                    encoder_name, decoder_name, config=config
+                )
+                self.encoder_tokenizer = tokenizer_class.from_pretrained(encoder_name)
+                self.decoder_tokenizer = BertTokenizer.from_pretrained(decoder_name)
+            self.encoder_config = self.model.config.encoder
+            self.decoder_config = self.model.config.decoder
 
         if self.args["wandb_project"] and not wandb_available:
             warnings.warn("wandb_project specified but wandb is not available. Wandb disabled.")
             self.args["wandb_project"] = None
+
+        if encoder_decoder_name:
+            self.args["model_name"] = encoder_decoder_name
+
+            # Checking if we are loading from a saved model or using a pre-trained model
+            if not saved_model_args and encoder_decoder_type == "marian":
+                # Need to store base pre-trained model name to get the tokenizer when loading a saved model
+                self.args["base_marian_model_name"] = encoder_decoder_name
+
+        elif encoder_name and decoder_name:
+            self.args["model_name"] = encoder_name + "-" + decoder_name
+        else:
+            self.args["model_name"] = "encoder-decoder"
+
+        if encoder_decoder_type:
+            self.args["model_type"] = encoder_decoder_type
+        elif encoder_type:
+            self.args["model_type"] = encoder_type + "-bert"
+        else:
+            self.args["model_type"] = "encoder-decoder"
 
     def train_model(
         self, train_data, output_dir=None, show_running_loss=True, args=None, eval_data=None, verbose=True, **kwargs,
@@ -122,10 +231,9 @@ class T5Model:
         Trains the model using 'train_data'
 
         Args:
-            train_data: Pandas DataFrame containing the 3 columns - `prefix`, `input_text`, `target_text`.
-                        - `prefix`: A string indicating the task to perform. (E.g. `"question"`, `"stsb"`)
-                        - `input_text`: The input text sequence. `prefix` is automatically prepended to form the full input. (<prefix>: <input_text>)
-                        - `target_text`: The target sequence
+            train_data: Pandas DataFrame containing the 2 columns - `input_text`, `target_text`.
+                        - `input_text`: The input text sequence.
+                        - `target_text`: The target sequence         
             output_dir: The directory where model files will be saved. If not given, self.args['output_dir'] will be used.
             show_running_loss (optional): Set to False to prevent running loss from being printed to console. Defaults to True.
             args (optional): Optional changes to the args dict of the model. Any changes made will persist for the model.
@@ -174,10 +282,13 @@ class T5Model:
             **kwargs,
         )
 
-        model_to_save = self.model.module if hasattr(self.model, "module") else self.model
-        model_to_save.save_pretrained(output_dir)
-        self.tokenizer.save_pretrained(output_dir)
-        torch.save(self.args, os.path.join(output_dir, "training_args.bin"))
+        self._save_model(self.args["output_dir"], model=self.model)
+
+        # model_to_save = self.model.module if hasattr(self.model, "module") else self.model
+        # model_to_save.save_pretrained(output_dir)
+        # self.encoder_tokenizer.save_pretrained(output_dir)
+        # self.decoder_tokenizer.save_pretrained(output_dir)
+        # torch.save(self.args, os.path.join(output_dir, "training_args.bin"))
 
         if verbose:
             logger.info(" Training of {} model complete. Saved to {}.".format(self.args["model_name"], output_dir))
@@ -193,7 +304,6 @@ class T5Model:
 
         model = self.model
         args = self.args
-        device = self.device
 
         tb_writer = SummaryWriter(logdir=args["tensorboard_dir"])
         train_sampler = RandomSampler(train_dataset)
@@ -219,6 +329,7 @@ class T5Model:
         warmup_steps = math.ceil(t_total * args["warmup_ratio"])
         args["warmup_steps"] = warmup_steps if args["warmup_steps"] == 0 else args["warmup_steps"]
 
+        # TODO: Use custom optimizer like with BertSum?
         optimizer = AdamW(optimizer_grouped_parameters, lr=args["learning_rate"], eps=args["adam_epsilon"])
         scheduler = get_linear_schedule_with_warmup(
             optimizer, num_warmup_steps=args["warmup_steps"], num_training_steps=t_total
@@ -294,7 +405,7 @@ class T5Model:
                 if steps_trained_in_current_epoch > 0:
                     steps_trained_in_current_epoch -= 1
                     continue
-                batch = tuple(t.to(device) for t in batch)
+                # batch = tuple(t.to(device) for t in batch)
 
                 inputs = self._get_inputs_dict(batch)
                 outputs = model(**inputs)
@@ -389,18 +500,20 @@ class T5Model:
 
                         if not best_eval_metric:
                             best_eval_metric = results[args["early_stopping_metric"]]
-                            self._save_model(
-                                args["best_model_dir"], optimizer, scheduler, model=model, results=results
-                            )
+                            if args["save_best_model"]:
+                                self._save_model(
+                                    args["best_model_dir"], optimizer, scheduler, model=model, results=results
+                                )
                         if best_eval_metric and args["early_stopping_metric_minimize"]:
                             if (
                                 results[args["early_stopping_metric"]] - best_eval_metric
                                 < args["early_stopping_delta"]
                             ):
                                 best_eval_metric = results[args["early_stopping_metric"]]
-                                self._save_model(
-                                    args["best_model_dir"], optimizer, scheduler, model=model, results=results
-                                )
+                                if args["save_best_model"]:
+                                    self._save_model(
+                                        args["best_model_dir"], optimizer, scheduler, model=model, results=results
+                                    )
                                 early_stopping_counter = 0
                             else:
                                 if args["use_early_stopping"]:
@@ -424,9 +537,10 @@ class T5Model:
                                 > args["early_stopping_delta"]
                             ):
                                 best_eval_metric = results[args["early_stopping_metric"]]
-                                self._save_model(
-                                    args["best_model_dir"], optimizer, scheduler, model=model, results=results
-                                )
+                                if args["save_best_model"]:
+                                    self._save_model(
+                                        args["best_model_dir"], optimizer, scheduler, model=model, results=results
+                                    )
                                 early_stopping_counter = 0
                             else:
                                 if args["use_early_stopping"]:
@@ -459,7 +573,8 @@ class T5Model:
                     eval_data, verbose=verbose and args["evaluate_during_training_verbose"], silent=True, **kwargs
                 )
 
-                self._save_model(output_dir_current, optimizer, scheduler, results=results)
+                if args["save_eval_checkpoints"]:
+                    self._save_model(output_dir_current, optimizer, scheduler, results=results)
 
                 training_progress_scores["global_step"].append(global_step)
                 training_progress_scores["train_loss"].append(current_loss)
@@ -473,11 +588,15 @@ class T5Model:
 
                 if not best_eval_metric:
                     best_eval_metric = results[args["early_stopping_metric"]]
-                    self._save_model(args["best_model_dir"], optimizer, scheduler, model=model, results=results)
+                    if args["save_best_model"]:
+                        self._save_model(args["best_model_dir"], optimizer, scheduler, model=model, results=results)
                 if best_eval_metric and args["early_stopping_metric_minimize"]:
                     if results[args["early_stopping_metric"]] - best_eval_metric < args["early_stopping_delta"]:
                         best_eval_metric = results[args["early_stopping_metric"]]
-                        self._save_model(args["best_model_dir"], optimizer, scheduler, model=model, results=results)
+                        if args["save_best_model"]:
+                            self._save_model(
+                                args["best_model_dir"], optimizer, scheduler, model=model, results=results
+                            )
                         early_stopping_counter = 0
                     else:
                         if args["use_early_stopping"] and args["early_stopping_consider_epochs"]:
@@ -496,7 +615,10 @@ class T5Model:
                 else:
                     if results[args["early_stopping_metric"]] - best_eval_metric > args["early_stopping_delta"]:
                         best_eval_metric = results[args["early_stopping_metric"]]
-                        self._save_model(args["best_model_dir"], optimizer, scheduler, model=model, results=results)
+                        if args["save_best_model"]:
+                            self._save_model(
+                                args["best_model_dir"], optimizer, scheduler, model=model, results=results
+                            )
                         early_stopping_counter = 0
                     else:
                         if args["use_early_stopping"] and args["early_stopping_consider_epochs"]:
@@ -520,9 +642,8 @@ class T5Model:
         Evaluates the model on eval_data. Saves results to output_dir.
 
         Args:
-            eval_data: Pandas DataFrame containing the 3 columns - `prefix`, `input_text`, `target_text`.
-                        - `prefix`: A string indicating the task to perform. (E.g. `"question"`, `"stsb"`)
-                        - `input_text`: The input text sequence. `prefix` is automatically prepended to form the full input. (<prefix>: <input_text>)
+            eval_data: Pandas DataFrame containing the 2 columns - `input_text`, `target_text`.
+                        - `input_text`: The input text sequence.
                         - `target_text`: The target sequence            
             output_dir: The directory where model files will be saved. If not given, self.args['output_dir'] will be used.
             verbose: If verbose, results will be printed to the console on completion of evaluation.
@@ -546,10 +667,7 @@ class T5Model:
         self.results.update(result)
 
         if self.args["evaluate_generated_text"]:
-            to_predict = [
-                prefix + ": " + input_text + " </s>"
-                for prefix, input_text in zip(eval_data["prefix"], eval_data["input_text"])
-            ]
+            to_predict = eval_data["input_text"].tolist()
             preds = self.predict(to_predict)
 
             result = self.compute_metrics(eval_data["target_text"].tolist(), preds, **kwargs)
@@ -570,7 +688,6 @@ class T5Model:
         model = self.model
         args = self.args
         eval_output_dir = output_dir
-        device = self.device
 
         results = {}
 
@@ -585,7 +702,7 @@ class T5Model:
         model.eval()
 
         for batch in tqdm(eval_dataloader, disable=args["silent"] or silent):
-            batch = tuple(t.to(device) for t in batch)
+            # batch = tuple(t.to(device) for t in batch)
 
             inputs = self._get_inputs_dict(batch)
             with torch.no_grad():
@@ -624,31 +741,37 @@ class T5Model:
             to_predict[i : i + self.args["eval_batch_size"]]
             for i in range(0, len(to_predict), self.args["eval_batch_size"])
         ]:
-            if self.args["preprocess_inputs"]:
-                input_ids = self.tokenizer.batch_encode_plus(
-                    [t + " </s>" for t in batch],
-                    max_length=self.args["max_seq_length"],
-                    pad_to_max_length=True,
-                    return_tensors="pt",
-                )["input_ids"]
-            else:
-                input_ids = self.tokenizer.batch_encode_plus(
-                    batch, max_length=self.args["max_seq_length"], pad_to_max_length=True, return_tensors="pt",
-                )["input_ids"]
+            input_ids = self.encoder_tokenizer.batch_encode_plus(
+                batch, max_length=self.args["max_seq_length"], pad_to_max_length=True, return_tensors="pt",
+            )["input_ids"]
             input_ids = input_ids.to(self.device)
-            outputs = self.model.generate(
-                input_ids=input_ids,
-                num_beams=self.args["num_beams"],
-                max_length=self.args["max_length"],
-                length_penalty=self.args["length_penalty"],
-                early_stopping=self.args["early_stopping"],
-                repetition_penalty=self.args["repetition_penalty"],
-                do_sample=self.args["do_sample"],
-            )
+
+            if self.args["model_type"] in ["bart", "marian"]:
+                outputs = self.model.generate(
+                    input_ids=input_ids,
+                    num_beams=self.args["num_beams"],
+                    max_length=self.args["max_length"],
+                    length_penalty=self.args["length_penalty"],
+                    early_stopping=self.args["early_stopping"],
+                    repetition_penalty=self.args["repetition_penalty"],
+                    do_sample=self.args["do_sample"],
+                )
+            else:
+                outputs = self.model.generate(
+                    input_ids=input_ids,
+                    decoder_start_token_id=self.model.config.decoder.pad_token_id,
+                    num_beams=self.args["num_beams"],
+                    max_length=self.args["max_length"],
+                    length_penalty=self.args["length_penalty"],
+                    early_stopping=self.args["early_stopping"],
+                    repetition_penalty=self.args["repetition_penalty"],
+                    do_sample=self.args["do_sample"],
+                )
+
             all_outputs.extend(outputs)
 
         return [
-            self.tokenizer.decode(output_id, skip_special_tokens=True, clean_up_tokenization_spaces=True)
+            self.decoder_tokenizer.decode(output_id, skip_special_tokens=True, clean_up_tokenization_spaces=True)
             for output_id in all_outputs
         ]
 
@@ -674,17 +797,6 @@ class T5Model:
 
         return results
 
-    def _move_model_to_device(self):
-        self.model.to(self.device)
-
-    def _get_inputs_dict(self, batch):
-        lm_labels = batch[1]
-        lm_labels[lm_labels == self.tokenizer.pad_token_id] = -100
-
-        inputs = {"input_ids": batch[0], "lm_labels": lm_labels}
-
-        return inputs
-
     def load_and_cache_examples(self, data, evaluate=False, no_cache=False, verbose=True, silent=False):
         """
         Creates a T5Dataset from data.
@@ -692,7 +804,8 @@ class T5Model:
         Utility function for train() and eval() methods. Not intended to be used directly.
         """
 
-        tokenizer = self.tokenizer
+        encoder_tokenizer = self.encoder_tokenizer
+        decoder_tokenizer = self.decoder_tokenizer
         args = self.args
 
         if not no_cache:
@@ -704,11 +817,12 @@ class T5Model:
 
         if args["dataset_class"]:
             CustomDataset = args["dataset_class"]
-            return CustomDataset(tokenizer, args, data, mode)
+            return CustomDataset(encoder_tokenizer, decoder_tokenizer, args, data, mode)
         else:
-            return T5Dataset(tokenizer, self.args, data, mode,)
-
-        return T5Dataset
+            if args["model_type"] in ["bart", "marian"]:
+                return SimpleSummarizationDataset(encoder_tokenizer, self.args, data, mode)
+            else:
+                return Seq2SeqDataset(encoder_tokenizer, decoder_tokenizer, self.args, data, mode,)
 
     def _create_training_progress_scores(self, **kwargs):
         extra_metrics = {key: [] for key in kwargs}
@@ -724,24 +838,84 @@ class T5Model:
     def _get_last_metrics(self, metric_values):
         return {metric: values[-1] for metric, values in metric_values.items()}
 
-    def _save_model(self, output_dir, optimizer, scheduler, model=None, results=None):
+    def _save_model(self, output_dir, optimizer=None, scheduler=None, model=None, results=None):
         os.makedirs(output_dir, exist_ok=True)
+
+        logger.info(f"Saving model into {output_dir}")
 
         if model:
             # Take care of distributed/parallel training
             model_to_save = model.module if hasattr(model, "module") else model
-            model_to_save.save_pretrained(output_dir)
-            self.tokenizer.save_pretrained(output_dir)
-            torch.save(self.args, os.path.join(output_dir, "training_args.bin"))
-            torch.save(optimizer.state_dict(), os.path.join(output_dir, "optimizer.pt"))
-            torch.save(scheduler.state_dict(), os.path.join(output_dir, "scheduler.pt"))
             self._save_model_args(output_dir)
+
+            if self.args["model_type"] in ["bart", "marian"]:
+                os.makedirs(os.path.join(output_dir), exist_ok=True)
+                model_to_save.save_pretrained(output_dir)
+                self.config.save_pretrained(output_dir)
+                if self.args["model_type"] == "bart":
+                    self.encoder_tokenizer.save_pretrained(output_dir)
+            else:
+                os.makedirs(os.path.join(output_dir, "encoder"), exist_ok=True)
+                os.makedirs(os.path.join(output_dir, "decoder"), exist_ok=True)
+                self.encoder_config.save_pretrained(os.path.join(output_dir, "encoder"))
+                self.decoder_config.save_pretrained(os.path.join(output_dir, "decoder"))
+
+                model_to_save = (
+                    self.model.encoder.module if hasattr(self.model.encoder, "module") else self.model.encoder
+                )
+                model_to_save.save_pretrained(os.path.join(output_dir, "encoder"))
+
+                model_to_save = (
+                    self.model.decoder.module if hasattr(self.model.decoder, "module") else self.model.decoder
+                )
+
+                model_to_save.save_pretrained(os.path.join(output_dir, "decoder"))
+
+                self.encoder_tokenizer.save_pretrained(os.path.join(output_dir, "encoder"))
+                self.decoder_tokenizer.save_pretrained(os.path.join(output_dir, "decoder"))
+
+            torch.save(self.args, os.path.join(output_dir, "training_args.bin"))
+            if optimizer:
+                torch.save(optimizer.state_dict(), os.path.join(output_dir, "optimizer.pt"))
+            if scheduler:
+                torch.save(scheduler.state_dict(), os.path.join(output_dir, "scheduler.pt"))
 
         if results:
             output_eval_file = os.path.join(output_dir, "eval_results.txt")
             with open(output_eval_file, "w") as writer:
                 for key in sorted(results.keys()):
                     writer.write("{} = {}\n".format(key, str(results[key])))
+
+    def _move_model_to_device(self):
+        self.model.to(self.device)
+
+    def _get_inputs_dict(self, batch):
+        device = self.device
+        if self.args["model_type"] in ["bart", "marian"]:
+            pad_token_id = self.encoder_tokenizer.pad_token_id
+            source_ids, source_mask, y = batch["source_ids"], batch["source_mask"], batch["target_ids"]
+            y_ids = y[:, :-1].contiguous()
+            lm_labels = y[:, 1:].clone()
+            lm_labels[y[:, 1:] == pad_token_id] = -100
+
+            inputs = {
+                "input_ids": source_ids.to(device),
+                "attention_mask": source_mask.to(device),
+                "decoder_input_ids": y_ids.to(device),
+                "lm_labels": lm_labels.to(device),
+            }
+        else:
+            lm_labels = batch[1]
+            lm_labels_masked = lm_labels.clone()
+            lm_labels_masked[lm_labels_masked == self.decoder_tokenizer.pad_token_id] = -100
+
+            inputs = {
+                "input_ids": batch[0].to(device),
+                "decoder_input_ids": lm_labels.to(device),
+                "lm_labels": lm_labels_masked.to(device),
+            }
+
+        return inputs
 
     def _save_model_args(self, output_dir):
         os.makedirs(output_dir, exist_ok=True)