+++ CHANGELOG.md
@@ -4,6 +4,42 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.28.0] - 2020-05-11
+
+### Added
+
+- Sequence-to-Sequence task support added. This includes the following models:
+  - BART
+  - Marian
+  - Generic Encoder-Decoder
+- The `args` dict of a task-specific Simple Transformers model is now saved along with the model. When loading the model, these values will be read and used.
+Any new `args` passed into the model initialization will override the loaded values.
+
+## [0.27.3] - 2020-05-10
+
+### Added
+
+- Support for `AutoModel` in NER, QA, and LanguageModeling. [@flozi00](https://github.com/flozi00)
+
+### Fixed
+
+- Now predict function from NER_Model returns a value model_outputs that contains:
+  A Python list of lists with dicts containing each word mapped to its list with raw model output. [@flaviussn](https://github.com/flaviussn)
+
+### Changed
+
+- Pillow import is now optional. It only needs to be installed if MultiModal models are used.
+
+## [0.27.2] - 2020-05-08
+
+### Fixed
+
+- Fixed T5 lm_labels not being masked properly
+
+### Changed
+
+- Torchvision import is now optional. It only needs to be installed if MultiModal models are used.
+
 ## [0.27.1] - 2020-05-05
 
 ### Fixed
@@ -647,7 +683,13 @@ Model checkpoint is now saved for all epochs again.
 
 - This CHANGELOG file to hopefully serve as an evolving example of a standardized open source project CHANGELOG.
 
-[0.27.1]: https://github.com/ThilinaRajapakse/simpletransformers/compare/32d5a1a...HEAD
+[0.28.0]: https://github.com/ThilinaRajapakse/simpletransformers/compare/402bd8e...HEAD
+
+[0.27.3]: https://github.com/ThilinaRajapakse/simpletransformers/compare/bc94b34...402bd8e
+
+[0.27.2]: https://github.com/ThilinaRajapakse/simpletransformers/compare/d665494...bc94b34
+
+[0.27.1]: https://github.com/ThilinaRajapakse/simpletransformers/compare/32d5a1a...d665494
 
 [0.27.0]: https://github.com/ThilinaRajapakse/simpletransformers/compare/ab1e600...32d5a1a